:hammer: 숙소 즐겨찾기 삭제시 userId 확인 추가

+++ backend/accommodation/src/main/java/leavelive/accommodation/service/AccommodationFavServiceImpl.java
@@ -8,6 +8,7 @@ import leavelive.accommodation.repository.AccommodationRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import javax.swing.text.html.Option;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
@@ -37,7 +38,10 @@ public class AccommodationFavServiceImpl {
         favRepo.save(entity.of(dto));
         return dto;
     }
-    public Long delete(Long id){
+    public Long delete(Long id,String userId){
+        Optional<AccommodationFav> accommodationFav=favRepo.findById(id);
+        if(!accommodationFav.isPresent()) throw new NullPointerException("해당하는 숙소가 없습니다.");
+        if(!accommodationFav.get().getUserId().equals(userId)) throw new NullPointerException("자신이 등록한 숙소만 삭제할 수 있습니다.");
         favRepo.deleteById(id);
         return id;
     }
